requirements.txt
azure-functions==1.13.1
openai
psycopg2-binary
langchain-community
langchain-openai
langchain-chroma
langchain
langchain-core
chromadb-client
azure-storage-blob
azure-identity
azure-keyvault-secrets
pypdf
cryptography==41.0.2